pyproject.toml
[project]
name = "pdf-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "camelot-py[cv]>=1.0.0",
    "ghostscript>=0.7"
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
